Add resolve.root option and update imports
This gets rid of all the `./` prefixes on the imports. I have left them
in the UI jsx files because maybe it makes sense to have the actions
subfolder always relative. Perhaps not. Perhaps the actions shouldn't
have a folder and should live alongside the component that 'owns' them
if such a component exists.